Update Public key constants as per FDO spec (#42)
- Update the used PublicKey.pkType constant for EC256 and EC284 key
types as FDO spec.

+++ crypto/openssl/openssl_RSAVerifyRoutines.c
@@ -182,7 +182,7 @@ err:
  * @param key_param2Length - size of public key2, type size_t
  * @return 0 if true, else -1.
  */
-int32_t crypto_hal_sig_verify(uint8_t key_encoding, uint8_t key_algorithm,
+int32_t crypto_hal_sig_verify(uint8_t key_encoding, int key_algorithm,
 			      const uint8_t *message, uint32_t message_length,
 			      const uint8_t *message_signature,
 			      uint32_t signature_length,
